+++ backend/app/Filament/Resources/CityResource/Pages/CreateCity.php
@@ -5,9 +5,19 @@ declare(strict_types=1);
 namespace App\Filament\Resources\CityResource\Pages;
 
 use App\Filament\Resources\CityResource;
+use Filament\Actions;
 use Filament\Resources\Pages\CreateRecord;
 
 class CreateCity extends CreateRecord
 {
+    use CreateRecord\Concerns\Translatable;
+
     protected static string $resource = CityResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            Actions\LocaleSwitcher::make(),
+        ];
+    }
 }