+++ backend/app/Filament/Resources/CityResource/Pages/EditCity.php
@@ -5,9 +5,21 @@ declare(strict_types=1);
 namespace App\Filament\Resources\CityResource\Pages;
 
 use App\Filament\Resources\CityResource;
+use Filament\Actions;
 use Filament\Resources\Pages\EditRecord;
 
 class EditCity extends EditRecord
 {
+    use EditRecord\Concerns\Translatable;
+
     protected static string $resource = CityResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            Actions\LocaleSwitcher::make(),
+            Actions\ViewAction::make(),
+            Actions\DeleteAction::make(),
+        ];
+    }
 }