+++ backend/app/Filament/Resources/CityResource/Pages/ListCities.php
@@ -10,11 +10,14 @@ use Filament\Resources\Pages\ListRecords;
 
 class ListCities extends ListRecords
 {
+    use ListRecords\Concerns\Translatable;
+
     protected static string $resource = CityResource::class;
 
     protected function getHeaderActions(): array
     {
         return [
+            Actions\LocaleSwitcher::make(),
             Actions\CreateAction::make(),
         ];
     }