+++ backend/app/Filament/Resources/PetTypeResource/Pages/CreatePetType.php
@@ -5,11 +5,21 @@ declare(strict_types=1);
 namespace App\Filament\Resources\PetTypeResource\Pages;
 
 use App\Filament\Resources\PetTypeResource;
+use Filament\Actions;
 use Filament\Resources\Pages\CreateRecord;
 
 class CreatePetType extends CreateRecord
 {
+    use CreateRecord\Concerns\Translatable;
+
     protected static string $resource = PetTypeResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            Actions\LocaleSwitcher::make(),
+        ];
+    }
 
     protected function getRedirectUrl(): string
     {