+++ backend/app/Filament/Resources/PetTypeResource/Pages/EditPetType.php
@@ -12,11 +12,14 @@ use Filament\Resources\Pages\EditRecord;
 
 class EditPetType extends EditRecord
 {
+    use EditRecord\Concerns\Translatable;
+
     protected static string $resource = PetTypeResource::class;
 
     protected function getHeaderActions(): array
     {
         return [
+            Actions\LocaleSwitcher::make(),
             Actions\ViewAction::make(),
             Actions\DeleteAction::make()
                 ->before(function ($record) {