+++ backend/app/Filament/Resources/PetTypeResource/Pages/ViewPetType.php
@@ -10,11 +10,14 @@ use Filament\Resources\Pages\ViewRecord;
 
 class ViewPetType extends ViewRecord
 {
+    use ViewRecord\Concerns\Translatable;
+
     protected static string $resource = PetTypeResource::class;
 
     protected function getHeaderActions(): array
     {
         return [
+            Actions\LocaleSwitcher::make(),
             Actions\EditAction::make(),
         ];
     }